[norun] Change to no-nominal builtin types
In this change set, we have decided that it's time to introduce

  - tuple types
  - nominal integer types definable in café code, not in the language

and to remove `Int32` and `Unit` (replaced by a library type and empty tuple
respectively).

This commit set is not done - this is only the initial change, enough work to
get it compiling. It does not compile any real café code.